+++ packages/react-sdk/src/server/index.ts
@@ -1,0 +1,21 @@
+export type {
+  PantheonClientConfig,
+  PantheonAPIOptions,
+} from "@pantheon-systems/pcc-sdk-core";
+export { PantheonClient } from "@pantheon-systems/pcc-sdk-core";
+
+export {
+  getArticles,
+  getPaginatedArticles,
+  getRecommendedArticles,
+  getArticle,
+  getAllTags,
+  getArticleBySlugOrId,
+  PCCConvenienceFunctions,
+  updateConfig,
+} from "@pantheon-systems/pcc-sdk-core";
+export * from "@pantheon-systems/pcc-sdk-core/types";
+
+export * from "../core/pantheon-context";
+
+export { NextPantheonAPI as PantheonAPI } from "../core/pantheon-api";