+++ packages/browser/CHANGELOG.md
@@ -1,5 +1,45 @@
 # @pantheon-systems/pcc-browser-sdk
 
+## 3.5.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @pantheon-systems/pcc-sdk-core@3.5.3
+
+## 3.5.3-beta.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @pantheon-systems/pcc-sdk-core@3.5.3-beta.0
+
+## 3.5.2
+
+### Patch Changes
+
+- 289e05c: Add updateConfig function which will override the default PCC
+  connection settings that PCCConvenienceFunctions will use.
+- Updated dependencies [fcfa574]
+- Updated dependencies [289e05c]
+  - @pantheon-systems/pcc-sdk-core@3.5.2
+
+## 3.5.2-beta.1
+
+### Patch Changes
+
+- Add updateConfig function which will override the default PCC connection
+  settings that PCCConvenienceFunctions will use.
+- Updated dependencies
+  - @pantheon-systems/pcc-sdk-core@3.5.2-beta.1
+
+## 3.5.2-beta.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @pantheon-systems/pcc-sdk-core@3.5.2-beta.0
+
 ## 3.5.1
 
 ### Patch Changes